% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda_gird.R
\name{Lambda_grid}
\alias{Lambda_grid}
\title{Simulate a grid set of lambdas for a given alpha in penalized regression.}
\usage{
Lambda_grid(X, y, n_lambda, alpha, loss_func)
}
\arguments{
\item{X}{expression levels of n_genes target genes (TGs).}

\item{y}{expression levels of a transcription factor (TF).}

\item{n_lambda}{the number of lambdas.}

\item{alpha}{the proportion of l1 norm affects (the numerical values of nonzero coefficients). It's in range (0,1].}

\item{loss_func}{either "Huber" or "MSE".}
}
\value{
n_lambda length vector of lambdas according to the alpha you provided.
}
\description{
Simulate a grid set of lambdas for a given alpha in penalized regression.
}
